requirements.txt
affine==2.4.0
annotated-types==0.6.0
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
bleach==6.1.0
bokeh==3.3.4
boto3==1.34.54
botocore==1.34.54
Brotli==1.1.0
cdsapi==0.6.1
certifi==2024.2.2
cffi==1.16.0
cfgrib==0.9.10.4
cfgv==3.3.1
cftime==1.6.3
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
cytoolz==0.12.3
dask==2024.2.1
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.8
distributed==2024.2.1
docutils==0.20.1
eccodes==1.7.0
entrypoints==0.4
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
findlibs==0.0.5
fonttools==4.49.0
fsspec==2024.2.0
identify==2.5.35
idna==3.6
importlib-metadata==7.0.1
importlib_resources==6.1.2
ipykernel==6.29.3
ipython==8.22.1
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyterlab_pygments==0.3.0
kiwisolver==1.4.5
locket==1.0.0
lockfile==0.12.2
luigi==3.5.0
lz4==4.3.3
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mistune==3.0.2
msgpack==1.0.7
munkres==1.1.4
nbclient==0.8.0
nbconvert==7.16.1
nbformat==5.9.2
nest_asyncio==1.6.0
netCDF4==1.6.5
nodeenv==1.8.0
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pip==24.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
plottable==0.1.5
ply==3.11
pre_commit==3.6.2
prometheus-client==0.14.1
prompt-toolkit==3.0.42
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyarrow-hotfix==0.6
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
Pygments==2.17.2
pyparsing==3.1.1
pyproj==3.6.1
PyQt5==5.15.9
PyQt5-sip==12.12.2
PySocks==1.7.1
python-daemon==3.0.1
python-dateutil==2.9.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
rasterio==1.3.9
referencing==0.33.0
requests==2.31.0
rioxarray==0.15.1
rpds-py==0.18.0
s3transfer==0.10.0
scipy==1.12.0
setuptools==69.1.1
sip==6.7.12
six==1.16.0
snuggs==1.4.7
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.2
tblib==3.0.0
tenacity==8.2.3
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
typing_extensions==4.10.0
tzdata==2024.1
ukkonen==1.0.1
urllib3==2.0.7
virtualenv==20.25.1
wcwidth==0.2.13
webencodings==0.5.1
wheel==0.42.0
xarray==2024.2.0
xyzservices==2023.10.1
zict==3.0.0
zipp==3.17.0
